+++ XHuiCloud-business/XHuiCloud-push/XHuiCloud-push-api/src/main/java/com/xhuicloud/push/common/PushMultiple.java
@@ -42,7 +42,7 @@ public class PushMultiple extends BasePushData{
 
     @ApiModelProperty(value = "用户id")
     @NotNull
-    private List<Integer> userIds;
+    private List<Long> userIds;
 
 }
