+++ XHuiCloud-commons/XHuiCloud-common-core/src/main/java/com/xhuicloud/common/core/constant/CommonConstants.java
@@ -105,7 +105,7 @@ public interface CommonConstants {
     /**
      * 默认租户ID
      */
-    Integer DEFAULT_TENANT_ID = 1;
+    Long DEFAULT_TENANT_ID = 1L;
 
     /**
      * 支付宝 ua
