+++ XHuiCloud-commons/XHuiCloud-common-feign/src/main/java/com/xhuicloud/common/feign/interceptor/XHuiFeignTenantInterceptor.java
@@ -42,7 +42,7 @@ public class XHuiFeignTenantInterceptor implements RequestInterceptor {
 
     @Override
     public void apply(RequestTemplate requestTemplate) {
-        Integer tenant = XHuiCommonThreadLocalHolder.getTenant();
+        Long tenant = XHuiCommonThreadLocalHolder.getTenant();
         if (tenant != null) {
             requestTemplate.header(CommonConstants.TENANT_ID, tenant.toString());
         }
