+++ XHuiCloud-upms/XHuiCloud-upms-api/src/main/java/com/xhuicloud/upms/entity/SysUserRole.java
@@ -28,19 +28,17 @@ import com.baomidou.mybatisplus.extension.activerecord.Model;
 import io.swagger.annotations.ApiModel;
 import lombok.Data;
 
-import java.io.Serializable;
-
 @Data
 @ApiModel(value="用户角色")
 public class SysUserRole extends Model<SysUserRole> {
     /**
      * 用户id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 角色id
      */
-    private Integer roleId;
+    private Long roleId;
 
 }
